Adding some logging and fixing things

+++ influxdbv2/resource_create_buckets.go
@@ -161,7 +161,7 @@ func getRetentionRules(input interface{}) (domain.RetentionRules, error) {
 	for _, retentionRule := range retentionRulesSet {
 		rr, ok := retentionRule.(map[string]interface{})
 		if ok {
-			each := domain.RetentionRule{EverySeconds: rr["every_seconds"].(int)}
+			each := domain.RetentionRule{EverySeconds: rr["every_seconds"].(int64)}
 			result = append(result, each)
 		}
 	}
